Rearranged packages and classes based on the MuleSoft checklist
Also made a minor change to the POM
Also changed the tab location for one connection parameter (Hostname Verifier).
This should cover all of the required changes for "Best Practices - Connector Interface", except for the single row which we intend to push back on (output schema).
