+++ pom.xml
@@ -6,9 +6,9 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.marklogic</groupId>
   <artifactId>marklogic-mule-connector</artifactId>
-  <version>2.0.0-SNAPSHOT</version>
+  <version>2.0.1-SNAPSHOT</version>
   <packaging>mule-extension</packaging>
-  <name>MarkLogic Connector for Mule</name>
+  <name>MarkLogic Connector for Mule 4</name>
 
   <parent>
     <groupId>org.mule.extensions</groupId>
