+++ src/main/java/com/marklogic/mule/connector/api/QueryParameters.java
@@ -1,4 +1,4 @@
-package com.marklogic.mule.internal;
+package com.marklogic.mule.connector.api;
 
 import com.marklogic.client.DatabaseClient;
 import com.marklogic.client.document.DocumentManager;
@@ -6,8 +6,9 @@ import com.marklogic.client.io.Format;
 import com.marklogic.client.io.StringHandle;
 import com.marklogic.client.query.QueryDefinition;
 import com.marklogic.client.query.QueryManager;
-import com.marklogic.mule.internal.api.QueryFormat;
-import com.marklogic.mule.internal.api.QueryType;
+import com.marklogic.mule.connector.internal.Utilities;
+import com.marklogic.mule.connector.internal.provider.QueryFormat;
+import com.marklogic.mule.connector.internal.provider.QueryType;
 import org.mule.runtime.extension.api.annotation.param.NullSafe;
 import org.mule.runtime.extension.api.annotation.param.Optional;
 import org.mule.runtime.extension.api.annotation.param.Parameter;
