+++ src/main/java/com/marklogic/mule/connector/api/extension/MarkLogicConnector.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.marklogic.mule.internal.extension;
+package com.marklogic.mule.connector.api.extension;
 
-import com.marklogic.mule.internal.config.Configuration;
-import com.marklogic.mule.internal.error.ErrorType;
+import com.marklogic.mule.connector.api.config.Configuration;
+import com.marklogic.mule.connector.internal.error.ErrorType;
 import org.mule.runtime.api.meta.Category;
 import org.mule.runtime.extension.api.annotation.Configurations;
 import org.mule.runtime.extension.api.annotation.Extension;
