+++ src/test/java/com/marklogic/mule/extension/DocumentData.java
@@ -1,6 +1,6 @@
 package com.marklogic.mule.extension;
 
-import com.marklogic.mule.internal.api.DocumentAttributes;
+import com.marklogic.mule.connector.internal.provider.DocumentAttributes;
 
 public class DocumentData {
